Add 'chalice dev plan' command
This introduces a new top level 'dev' namespace for chalice commands.
This is where we can add helpful commands to introspect the internals
of the chalice deployment process.  I've updated the docs to make it
clear that there is no expectation of backwards compatibility with
these commands, given that they're surfacing internals.

The first of these commands is the 'plan' command which calculates,
but does not execute the chalice plan.  It also pretty prints it
so it's easier to read.