Preserve file permissions in zipfile generation
This removes the code to always use 0o644 for all files in the zip.
We want to preserve executable files.  I think there's some additional
work where we can always ensure that files in the zip are readable.
This would fix the issue where if the file isn't readable for whatever
reason, lambda will fail in ways that's confusing to users.  However,
I think that's a separate issue from this.