Rework reloading to be simpler
The basic functionality works for the stat-based poller,
haven't looked at the watchdog stuff yet.

It seems like a process/forking server isn't a clear winner.
There's some tradeoffs that we have to consider.  I still think
in the long run we should probably go with this approach,
but there's still more work to be done.