Add fallback to check pyproject.toml for name/version
When using a git repo dependency with only a pyproject.toml,
we were failing with an UnsupportedPackageError.  Modern
projects *only* have a pyproject.toml file and no setup.py
so we need to handle this case now.

This lets you reference private python depdencies hosted
in a git repo that have no setup.py (and only pyproject.toml).